Improved paths concatenation
The __correctPath method has been rewritten as _concatenatePaths and takes any number of arguments.
The paths are now concatenated, the result is prefixed by a slash, and finally every multiple slashes are replaced by a single slash.
The main impact is that it is now possible to skip the prefix slashes in the paths.