Added some regression test - disabled
When a route that does not match because of some readOnly restrictions will cause a fallback to a default route, it will first impact the state. And some parts of it that are not in the default route restrictions will not be restored.
I think the state should actually not be affected at all...

Edit: The test has been disabled, because baobab-router cannot give the expected result in its current implementation.
Since the usecase is somehow a problem, I prefer to let it here, because it represent pretty well the mess it can cause.
Basically, when the router will be able to solve the stable pair state / hash without having to actually impact them until it's stable, then this case will no longer be a problem.